fix ai security warnning

Created-by: xia-bubai
Commit-by: xia-bubai
Merged-by: openharmony_ci
Description: ### 关联的issue：
https://gitcode.com/openharmony/security_access_token/issues/3551

1、TDD：
![image.png](https://raw.gitcode.com/user-images/assets/4390047/67b6a37f-495d-4bc9-a3e3-c5aa3788a727/image.png 'image.png')
2、XTS:

3、手工用例（自验证步骤/预期结果/实际结果）:

- [x] 不涉及API功能发生变化
- [x] 不涉及改变回调函数或生命周期的触发时机或时序，或删除生命周期和回调函数
- [x] 不涉及参数的规格发生了变化，例如取值范围缩小等
- [x] 不涉及对外接口新增权限校验
- [x] 不涉及接口的使用约束规格收紧，例如权限开放范围变化
- [x] 不涉及系统可创建的实例数量收紧
- [x] 不涉及修改了接口返回的数据
- [x] 未新增错误抛出，包括新增错误码、对已有的场景从不会抛错误码变成会抛出错误
- [x] 未修改已有的错误码，对于相同的输入原本抛出A错误码变成了抛出B错误码
- [x] 接口性能未出现明显劣化

- [x] 是否需要合入权限定义
    - [x] 仅向系统服务开放的权限（availableType为SERVICE的权限）不允许合入global_system_resources仓
    - [x] 非SERVICE的权限与global_system_resources联合构建，即两笔pr关联同一个issue
    - [x] 填入必要参数name/grantMode/availableLevel/since/provisionEnable/distributedSceneEnable， 且与global_system_resources中声明（如果有）一致
    - [x] 声明权限生效的设备平台范围deviceTypes，以列表方式声明("deviceTypes" : [ "xxx", "xxx"])，权限在所有设备通用为"general"，非全平台生效按需填写，包括但不限于"phone"、"wearable"、"tablet"、"2in1"、"tv"、"car"

- [x] 裸指针避免通过隐式转换构造为sptr
- [x] json对象在取值之前必须先判断类型，避免类型不匹配
- [x] 序列化时必须对传入的数组大小进行校验，避免出现超大数组
- [x] 避免使用未明确位宽的整型，选择使用int8_t、uint8_t等类型
- [x] 外部传入的路径要做规范化校验，对路径中的.、..、../等特殊字符严格校验
- [x] 指针变量、表示资源描述符的变量、bool变量必须赋初值
- [x] readParcelable获取的对象使用前需要判空
- [x] 分配和释放内存的函数需要成对出现
- [x] 申请内存后异常退出前需要及时进行内存释放
- [x] 内存申请前必须对内存大小进行合法性校验
- [x] 内存分配后必须判断是否成功
- [x] 禁止打印文件路径、口令等敏感信息，如有需要，使用private修饰
- [x] 禁止打印内存地址
- [x] 整数之间运算时必须严格检查，确保不会出现溢出、反转、除0
- [x] 禁止对有符号整数进行位操作符运算
- [x] 循环次数如果收外部数据控制，需要检验其合法性
- [x] 必须检查安全函数的返回值，并进行正确处理

- [x] 【规则】高频代码的正常流程中禁止打印日志
- [x] 【规则】在基本不可能发生的点必须要打印日志
- [x] 【规则】事件记录的日志使用who do what 主谓宾的形式打印
- [x] 【规则】状态变化的日志打印使用state_name:s1->s2, reason:msg的形式打印
- [x] 【规则】参数值的日志打印使用name1=value1, name2=value2…的形式打印

See merge request: openharmony/security_access_token!4825

AI[0%] Human Fixed[0%] Human[0%] AI Adopted[0%]

+++ frameworks/common/include/data_validator.h
@@ -68,7 +68,7 @@ public:
     static bool IsAclExtendedMapSizeValid(const std::map<std::string, std::string>& aclExtendedMap);
     static bool IsAclExtendedMapContentValid(const std::string& permissionName, const std::string& value);
     static bool IsAppProvisionTypeValid(const std::string& appProvisionType);
-    static bool IsListSizeValid(uint32_t size);
+    static bool IsListSizeValid(size_t size);
     static bool IsPermFeatureValid(const std::string& feature);
 
 private:
